# website/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Indexarr",
    "version": "0.1.0",
    "description": "Decentralized torrent indexing with DHT crawling, content classification, gossip-based P2P sync, and contributor identity.\n\n## Authentication\n\nMost endpoints require no authentication. Admin endpoints (API key management, nuke review, debug logging) require an API key passed via the `X-Api-Key` header. The Torznab endpoint supports an optional `apikey` query parameter.\n\n## Base URL\n\nAll REST endpoints are prefixed with `/api/v1/`. The Torznab endpoint is at `/api/torznab`. GraphQL is at `/graphql`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/torrents/tags": {
      "post": {
        "tags": ["Torrents"],
        "summary": "Add tags",
        "description": "Add tags to one or more torrents.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "info_hashes": { "type": "array", "items": { "type": "string" } },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Tags added" } }
      },
      "delete": {
        "tags": ["Torrents"],
        "summary": "Remove tags",
        "description": "Remove tags from one or more torrents.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "info_hashes": { "type": "array", "items": { "type": "string" } },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Tags removed" } }
      }
    },
    "/api/v1/torrents": {
      "delete": {
        "tags": ["Torrents"],
        "summary": "Delete torrents",
        "parameters": [
          { "name": "info_hashes", "in": "query", "schema": { "type": "array", "items": { "type": "string" } }, "description": "Info hashes to delete" }
        ],
        "responses": { "200": { "description": "Torrents deleted" } }
      }
    },
    "/api/v1/import": {
      "post": {
        "tags": ["Torrents"],
        "summary": "Bulk import",
        "description": "Import an array of torrent records.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "info_hash": { "type": "string" },
                    "name": { "type": "string" },
                    "size": { "type": "integer" },
                    "source": { "type": "string" },
                    "files": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Import results" } }
      }
    },
...
  }
}